twitter/requirements.txt
python-dotenv == 0.21.1
tweepy == 4.12.1
